Add vl6 score and debug efforts to understand zero results
Add the results from the vl6 run and produce scores as for hello world.
Add new code to generate challenge pids in the the original noholdout
to allow results to be matched on challenge pid and not just sort order.
Create new routine to score challenge on data set.
Debug challenge results for vl6.