Add fast load of mpd slice files using array plus concat method
Add new function for fast loading of the mpd slice files.
This loads the files in linear time because the final data frame
is build just once from pd.concat()

Update the call to use fast function.
Add the reindex into the original load function and avoid the
clunky index assignment for the playlist df.