Convert to challenge-to-all similarity computation
Avoid computing all-to-all similarity because it isn't need
and costs too much compute and ram.
This change also increases the accuracy of u2u results, presumably
by not polluting results with challenge participants (other queries).
This commit applies the function defined and called in commit 375de133
without the extraneous edits in that commit.

Introduce trainset_size and use it to select the appropriate parts
of the data sets for sim and score computation.
Add cosine_similaries2() to compute simlarities between
challenge and training set.
Adjust the score computation for the correct matrix shape.