Fix mfals wo transpose through upgrade to implicit 0.5.2
The earlier version of implicit 0.4.8 took an item-user matrix
to fit, hence the need to transpose.
The new version take user-item so my allmpb doesn't need to change.
The new version requires some additional changes to how the recommendations
are produced.
Add monitoring code to observe the shape of the user and item features after fit().
Add monitoring code to understand the results of recommend().

Create new kernel mpd3 for upgrades to run notebooks so update the notebook
runner kernel.