Add run harness for vl6 code and submit aicrowd-queue for mpd runs
Add the run-vl6 wrapper script to manage the runs of vl6 codes
external to this repo.
Add submit wrapper extensions to support selection of notebook algos
and vl6.
Add cache dir create in the results output to support downstream
scoring and results processing.
Update the notebook wrapper to support aicrowd-queue